+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/lib/ui/common/theme/theme.dart
@@ -12,31 +12,3 @@ final isDarkMode = RM.inject<bool>(
     // print(s);
   },
 );
-
-class _Theme {
-  static ThemeData get theme {
-    final themeData = ThemeData.dark();
-    final textTheme = themeData.textTheme;
-    final bodyText2 =
-        textTheme.bodyText2.copyWith(decorationColor: Colors.transparent);
-
-    return ThemeData.dark().copyWith(
-      primaryColor: Colors.grey[800],
-      accentColor: Colors.cyan[300],
-      buttonColor: Colors.grey[800],
-      textSelectionColor: Colors.cyan[100],
-      toggleableActiveColor: Colors.cyan[300],
-      floatingActionButtonTheme: FloatingActionButtonThemeData(
-        backgroundColor: Colors.cyan[300],
-      ),
-      snackBarTheme: SnackBarThemeData(
-        backgroundColor: themeData.dialogBackgroundColor,
-        contentTextStyle: bodyText2,
-        actionTextColor: Colors.cyan[300],
-      ),
-      textTheme: textTheme.copyWith(
-        bodyText2: bodyText2,
-      ),
-    );
-  }
-}
