+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/lib/ui/pages/add_edit_screen.dart/add_edit_screen.dart
@@ -4,13 +4,12 @@
 
 import 'dart:async';
 
+import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/injected/injected_todo.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
 
 import '../../../domain/entities/todo.dart';
-import '../../../injected.dart';
-import '../../../service/todos_state.dart';
 import '../../common/localization/localization.dart';
 
 class AddEditPage extends StatefulWidget {
