+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/lib/ui/pages/home_screen/extra_actions_button.dart
@@ -1,8 +1,9 @@
+import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/injected/injected_todo.dart';
+
+import '../../injected/injected_user.dart';
 import 'package:flutter/material.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
 
-import '../../../injected.dart';
-import '../../../service/todos_state.dart';
 import '../../common/enums.dart';
 import '../../common/localization/localization.dart';
 import '../../common/theme/theme.dart';
@@ -27,7 +28,7 @@ class ExtraActionsButton extends StatelessWidget {
           }
 
           if (action == ExtraAction.logout) {
-            user.state = authService.state.logout();
+            user.state = user.state.logout();
             return;
           }
