+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/lib/ui/pages/home_screen/home_screen.dart
@@ -1,7 +1,7 @@
+import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/injected/injected_todo.dart';
 import 'package:flutter/material.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
 
-import '../../../injected.dart';
 import '../../common/enums.dart';
 import '../../common/localization/localization.dart';
 import '../../pages/add_edit_screen.dart/add_edit_screen.dart';
