+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/lib/ui/pages/home_screen/stats_counter.dart
@@ -2,9 +2,9 @@
 // Use of this source code is governed by the MIT license that can be found
 // in the LICENSE file.
 
+import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/injected/injected_todo.dart';
 import 'package:flutter/material.dart';
 
-import '../../../injected.dart';
 import '../../../ui/common/localization/localization.dart';
 
 class StatsCounter extends StatelessWidget {
