+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/lib/ui/widgets/splash_screen.dart
@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
 
-import '../../injected.dart';
-
 class SplashScreen extends StatelessWidget {
   const SplashScreen([dynamic data]);
   static String routeName = '/';
   @override
   Widget build(BuildContext context) {
-    user.state;
     return Scaffold(
       body: Center(
         child: CircularProgressIndicator(),
