+++ examples/ex_009_1_4_ca_todo_mvc_with_state_persistence_and_user_auth/test/main_test.dart
@@ -1,3 +1,4 @@
+import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/injected/injected_user.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:states_rebuilder/states_rebuilder.dart';
@@ -6,15 +7,13 @@ import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/main.dar
 import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/pages/home_screen/extra_actions_button.dart';
 import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/ui/pages/home_screen/languages.dart';
 
-import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/injected.dart';
-
 import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/domain/common/extensions.dart';
 import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/domain/entities/user.dart';
 import 'package:ex_009_1_3_ca_todo_mvc_with_state_persistence_user_auth/domain/value_object/token.dart';
 import 'fake_auth_repository.dart';
 
 void main() async {
-  final storage = await RM.localStorageInitializerMock();
+  final storage = await RM.storageInitializerMock();
   authRepository.injectMock(() => FakeAuthRepository());
   setUp(() {
     storage.clear();
